moved AzureBlobStorageToGCSOperator to google provider
