+++ docs/apache-airflow-providers-google/operators/transfer/azure_blob_to_gcs.rst
@@ -18,34 +18,19 @@
 
 Azure Blob Storage to Google Cloud Storage (GCS) Transfer Operator
 ==================================================================
-The Blob service stores text and binary data as objects in the cloud.
-The Blob service offers the following three resources: the storage account, containers, and blobs.
-Within your storage account, containers provide a way to organize sets of blobs.
-For more information about the service visit `Azure Blob Storage API documentation <https://docs.microsoft.com/en-us/rest/api/storageservices/blob-service-rest-api>`_.
+The `Google Cloud Storage <https://cloud.google.com/storage/>`__  (GCS) is used to store large data from various applications.
+This is also the same with `Azure Blob Storage <https://docs.microsoft.com/en-us/rest/api/storageservices/blob-service-rest-api>`__.
+This page shows how to transfer data from Azure Blob Storage to GCS.
 
-Before you begin
-^^^^^^^^^^^^^^^^
-Before using Blob Storage within Airflow you need to authenticate your account with Token, Login and Password.
-Please follow Azure
-`instructions <https://docs.microsoft.com/en-us/azure/storage/common/storage-account-keys-manage?tabs=azure-portal>`_
-to do it.
+Prerequisite Tasks
+^^^^^^^^^^^^^^^^^^
 
-TOKEN should be added to the Connection in Airflow in JSON format, Login and Password as plain text.
-You can check `how to do such connection <https://airflow.apache.org/docs/apache-airflow/stable/howto/connection/index.html#editing-a-connection-with-the-ui>`_.
+.. include:: ../_partials/prerequisite_tasks.rst
 
-See following example.
-Set values for these fields:
+Use the :class:`~airflow.providers.google.cloud.transfers.azure_blob_to_gcs.AzureBlobStorageToGCSOperator`
+to transfer data from Azure Blob Storage to Google Cloud Storage.
 
-.. code-block::
-
-  Connection Id: wasb_default
-  Login: Storage Account Name
-  Password: KEY1
-  Extra: {"sas_token": "TOKEN"}
-
-.. _howto/operator:AzureBlobStorageToGCSOperator:
-
-Transfer Data from Blob Storage to Google Cloud Storage
+Transfer Data from Azure Blob Storage to Google Cloud Storage
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 Operator transfers data from Azure Blob Storage to specified bucket in Google Cloud Storage
 
@@ -54,7 +39,16 @@ To get information about jobs within a Azure Blob Storage use:
 
 Example usage:
 
-.. exampleinclude:: /../../tests/system/providers/microsoft/azure/example_azure_blob_to_gcs.py
+.. exampleinclude:: /../../tests/system/providers/google/cloud/example_azure_blob_to_gcs.py
     :language: python
     :start-after: [START how_to_azure_blob_to_gcs]
     :end-before: [END how_to_azure_blob_to_gcs]
+
+Reference
+^^^^^^^^^
+
+For further information, look at:
+
+* `GCS Client Library Documentation <https://googleapis.dev/python/storage/latest/index.html>`__
+* `GCS Product Documentation <https://cloud.google.com/storage/docs/>`__
+* `Azure Blob Storage Client Library Documentation <https://learn.microsoft.com/en-us/azure/storage/blobs/storage-quickstart-blobs-python>`__
